Some fixes to installing nexus from a resource
- move helper class to its own helper library module
- force specify Nexus download url
- move include_recipe outside of resource and into recipe
- remove unnecessary log[...] resource
- more fixes in the service_* resources
- vmoptions hash merge
