+++ resources/service_sysvinit.rb
@@ -16,8 +16,6 @@ action :start do
   create_init
 
   service "nexus3_#{new_resource.instance_name}" do
-    # TODO: define in helper (see Tomcat)
-    provider platform_sysv_init_class
     supports restart: true, status: true
     action :start
   end
@@ -25,16 +23,13 @@ end
 
 action :stop do
   service "nexus3_#{new_resource.instance_name}" do
-    provider platform_sysv_init_class
     supports status: true
     action :stop
-    only_if { ::File.exist?("/etc/init.d/nexus3_#{new_resource.instance_name}") }
   end
 end
 
 action :restart do
   service "nexus3_#{new_resource.instance_name}" do
-    provider platform_sysv_init_class
     supports status: true
     action :restart
   end
@@ -44,19 +39,15 @@ action :enable do
   create_init
 
   service "nexus3_#{new_resource.instance_name}" do
-    provider platform_sysv_init_class
     supports status: true
     action :enable
-    only_if { ::File.exist?("/etc/init.d/nexus3_#{new_resource.instance_name}") }
   end
 end
 
 action :disable do
   service "nexus3_#{new_resource.instance_name}" do
-    provider platform_sysv_init_class
     supports status: true
     action :disable
-    only_if { ::File.exist?("/etc/init.d/nexus3_#{new_resource.instance_name}") }
   end
 end
 
@@ -66,20 +57,6 @@ action_class do
       to ::File.join(new_resource.install_dir.to_s, 'bin', 'nexus')
       owner new_resource.nexus3_user
       group new_resource.nexus3_group
-    end
-
-    if platform_family?('rhel', 'fedora')
-      execute 'chkconfig add' do
-        command "chkconfig --add nexus3_#{new_resource.instance_name}"
-      end
-
-      execute 'chkconfig levels' do
-        command "chkconfig --levels 345 nexus3_#{new_resource.instance_name} on"
-      end
-    else
-      execute 'update-rc.d' do
-        command "update-rc.d nexus3_#{new_resource.instance_name} defaults"
-      end
     end
   end
