+++ spec/unit/resources/default_spec.rb
@@ -38,19 +38,25 @@ describe 'nexus3_test::default' do
     it 'creates rc template' do
       expect(chef_run).to create_template("/opt/nexus-#{VER}/bin/nexus.rc")
       expect(chef_run.template("/opt/nexus-#{VER}/bin/nexus.rc"))
-        .to notify('log[foo is restarting]').delayed
+        .to notify('nexus3_service[nexus3_foo]').to(:restart).delayed
+      expect(chef_run.template("/opt/nexus-#{VER}/bin/nexus.rc"))
+        .to notify('ruby_block[block until operational]').to(:create).delayed
     end
 
     it 'creates vmoptions template' do
       expect(chef_run).to create_template("/opt/nexus-#{VER}/bin/nexus.vmoptions")
       expect(chef_run.template("/opt/nexus-#{VER}/bin/nexus.vmoptions"))
-        .to notify('log[foo is restarting]').delayed
+        .to notify('nexus3_service[nexus3_foo]').to(:restart).delayed
+      expect(chef_run.template("/opt/nexus-#{VER}/bin/nexus.vmoptions"))
+        .to notify('ruby_block[block until operational]').to(:create).delayed
     end
 
     it 'creates properties template' do
       expect(chef_run).to create_template('/opt/sonatype-work/nexus3/etc/nexus.properties')
       expect(chef_run.template('/opt/sonatype-work/nexus3/etc/nexus.properties'))
-        .to notify('log[foo is restarting]').delayed
+        .to notify('nexus3_service[nexus3_foo]').to(:restart).delayed
+      expect(chef_run.template('/opt/sonatype-work/nexus3/etc/nexus.properties'))
+        .to notify('ruby_block[block until operational]').to(:create).delayed
     end
 
     it 'creates homedir symlink' do
@@ -58,11 +64,8 @@ describe 'nexus3_test::default' do
     end
 
     it 'enables nexus service' do
-      expect(chef_run).to enable_nexus3_service('foo')
-    end
-
-    it 'logs a service restart' do
-      expect(chef_run.log('foo is restarting')).to do_nothing
+      expect(chef_run).to enable_nexus3_service('nexus3_foo')
+      expect(chef_run).to enable_nexus3_service('nexus3_bar')
     end
 
     it 'blocks until operational does nothing' do
@@ -107,19 +110,25 @@ describe 'nexus3_test::default' do
     it 'creates rc template' do
       expect(chef_run).to create_template("C:\\Nexus3/nexus-#{VER}/bin/nexus.rc")
       expect(chef_run.template("C:\\Nexus3/nexus-#{VER}/bin/nexus.rc"))
-        .to notify('log[foo is restarting]').delayed
+        .to notify('nexus3_service[nexus3_foo]').to(:restart).delayed
+      expect(chef_run.template("C:\\Nexus3/nexus-#{VER}/bin/nexus.rc"))
+        .to notify('ruby_block[block until operational]').to(:create).delayed
     end
 
     it 'creates vmoptions template' do
       expect(chef_run).to create_template("C:\\Nexus3/nexus-#{VER}/bin/nexus.vmoptions")
       expect(chef_run.template("C:\\Nexus3/nexus-#{VER}/bin/nexus.vmoptions"))
-        .to notify('log[foo is restarting]').delayed
+        .to notify('nexus3_service[nexus3_foo]').to(:restart).delayed
+      expect(chef_run.template("C:\\Nexus3/nexus-#{VER}/bin/nexus.vmoptions"))
+        .to notify('ruby_block[block until operational]').to(:create).delayed
     end
 
     it 'creates properties template' do
       expect(chef_run).to create_template('C:\Nexus3/sonatype-work/nexus3/etc/nexus.properties')
       expect(chef_run.template('C:\Nexus3/sonatype-work/nexus3/etc/nexus.properties'))
-        .to notify('log[foo is restarting]').delayed
+        .to notify('nexus3_service[nexus3_foo]').to(:restart).delayed
+      expect(chef_run.template('C:\Nexus3/sonatype-work/nexus3/etc/nexus.properties'))
+        .to notify('ruby_block[block until operational]').to(:create).delayed
     end
 
     it 'creates homedir symlink' do
@@ -127,11 +136,7 @@ describe 'nexus3_test::default' do
     end
 
     it 'enables nexus service' do
-      expect(chef_run).to enable_nexus3_service('foo')
-    end
-
-    it 'logs a service restart' do
-      expect(chef_run.log('foo is restarting')).to do_nothing
+      expect(chef_run).to enable_nexus3_service('nexus3_foo')
     end
 
     it 'blocks until operational does nothing' do
