+++ test/fixtures/cookbooks/nexus3_test/recipes/default.rb
@@ -7,9 +7,7 @@ end
 
 # On Debian-style systems, curl is not installed on the kitchen
 # images. Install it here since we only need it for testing.
-package 'curl' do
-  only_if { platform_family?('debian') }
-end
+package 'curl'
 
 # Installation with default settings
 nexus3 'foo'
